+++ src/lib.rs
@@ -5,7 +5,9 @@ pub mod indicators;
 pub mod standard;
 
 use std::collections::HashMap;
-use std::io::BufRead;
+use std::fs::File;
+use std::io::{BufRead, Write};
+use std::path::PathBuf;
 
 use anyhow::Result;
 use log::warn;
@@ -28,6 +30,53 @@ macro_rules! add_indicators {
             }
         )*
     }
+}
+
+///
+/// # Errors
+///
+pub fn init(path: &PathBuf) -> std::io::Result<()> {
+    let mut file = File::create(path)?;
+
+    file.write_all(
+        b"\
+# `prepare` command
+#
+# Read the documentation at:
+# https://cardinal.readthedocs.io/en/latest/cli/prepare.html
+
+[defaults]
+# currency = USD
+# item_classification_scheme = UNSPSC
+# bid_status = valid
+# award_status = active
+
+[codelists.bidStatus]
+# qualified = valid
+
+# `indicators` command
+#
+# Read the documentation at:
+# https://cardinal.readthedocs.io/en/latest/cli/indicators/
+
+[R024]
+# threshold = 0.05
+
+[R025]
+# percentile = 75
+# threshold = 0.05
+
+[R035]
+# threshold = 1
+
+[R036]
+
+[R038]
+# threshold = 0.5
+",
+    )?;
+
+    Ok(())
 }
 
 fn fold_reduce<T: Send, Fold, Reduce, Finalize>(