+++ src/main.rs
@@ -59,6 +59,11 @@ enum Commands {
         #[arg(long, short, value_parser = settings_parser)]
         settings: Option<ocdscardinal::Settings>,
     },
+    /// Write a settings file for easy configuration.
+    Init {
+        /// The path to the settings file to write
+        file: PathBuf,
+    },
 }
 
 fn file_argument_error(file: &Path, message: &str) -> ! {
@@ -111,6 +116,13 @@ fn main() {
     pretty_env_logger::formatted_builder().filter_level(level).init();
 
     match &cli.command {
+        Commands::Init { file } => {
+            if let Err(err) = ocdscardinal::init(file) {
+                eprintln!("Error writing to {file:?}: {err}");
+            } else {
+                println!("Settings written to {file:?}.");
+            }
+        }
         Commands::Coverage { file } => match ocdscardinal::Coverage::run(reader(file)) {
             Ok(item) => {
                 println!("{:?}", item.results());